TerraPerkaCraft 2.0.0+1.20.1
TerraPerkaCraft 1.0.0+1.20.1 was tagged in 12e7c78, but I've since
decided to retcon the pack's version number in order to maintain a
linear Blolol TerraFirmaCraft mod pack history. As discussed on the wiki
at https://wiki.blolol.com/TerraPerkaCraft, "Season 1", and thus v1, of
TerraPerkaCraft is now considered to be the original Blolol TFC server
from 2012; and this iteration of the mod pack is now officially v2.