Perkpack 4: Ship default options.txt in Prism instance
YOSBR doesn't seem to be copying the default options.txt to the Minecraft
game root on first launch, perhaps because the game is generating the
file before YOSBR is instantiated. Instead, let's try shipping the file
in the Prism Launcher instance zip archive. That should work for this
file, because it's only needed on the client side. The other default
config files still need to be handled by YOSBR, because the server
doesn't use the Prism Launcher zip.